requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
numpy>=1.26.0
faiss-cpu>=1.7.4

